+++ sources/voting/urls.py
@@ -16,6 +16,4 @@ Including another URLconf
 from django.conf.urls import url
 from . import views
 
-urlpatterns = [
-    url(r'', views.index)
-]
+urlpatterns = []